Adapt UCO-Profile-Example to become CDO-Shapes-Example
This adaptation is made in partial satisfaction of ticket CDOTSC-34.

Licensing is carried forward from the "base" UCO Profile repository with
Paul's sign-off.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

References:
* [CDOTSC-34] CDO should provide shapes for external ontologies and
  other RDF models, including OWL

Signed-off-by: Paul Brandt <paul.brandt@politie.nl>
Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>