Review recursive path structures when assembling reviewed-properties set
This patch updates the loop that builds the set of properties involved
in shapes.  That set is tested against the properties used in the
exemplars graph to confirm exemplars exercise the properties in the
SHACL graph.

The `sh:alternativePath` gatherer loop did not address the recursive
potential of SHACL property shapes.  This patch adds a function to
recurse through the predicates available to property-paths, and more
generally enables doctests for the `pytest` script c/o this recursing
function having a doctest.

No effects were observed on Make-managed files.

References:
* https://www.w3.org/TR/2017/REC-shacl-20170720/#property-paths

Signed-off-by: Alex Nelsin <alexander.nelson@nist.gov>